EmployeeLeaveManagement/requirements.txt
psutil==5.4.8
Django==2.1.5
djangorestframework==3.9.0
mysqlclient==1.3.12
Pillow==5.3.0
django-widget-tweaks==1.4.3
django-model-utils==3.1.2
django-imagekit==4.0.2
django-debug-toolbar==1.11
python3-memcached==1.51
rpyc==4.0.2